refactor: app/util package 내에 용도를 쉽게 구별할 수 있도록 하위 package 생성
