+++ src/main/java/com/zelusik/eatery/app/config/WebMvcConfig.java
@@ -1,6 +1,6 @@
 package com.zelusik.eatery.app.config;
 
-import com.zelusik.eatery.app.util.ReviewKeywordRequestConverter;
+import com.zelusik.eatery.app.util.controller.ReviewKeywordRequestConverter;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.format.FormatterRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
