+++ src/main/java/com/zelusik/eatery/app/util/controller/ReviewKeywordRequestConverter.java
@@ -1,4 +1,4 @@
-package com.zelusik.eatery.app.util;
+package com.zelusik.eatery.app.util.controller;
 
 import com.zelusik.eatery.app.constant.review.ReviewKeyword;
 import org.springframework.core.convert.converter.Converter;
